Changing modules to public
Those modules need to be public to be able to use Verus as a library.
Also changed import_export.rs functions and structures to be public and
not package private.
Made one of the fields of verifier.rs public because it has no way to be
set other than using construct_vir() which is not a versitile function.